Better artifact handling
Return path to the chroot as an artifact. Wasteful to enumerate all the
files, the entire image, whether it's a mount point or just a directory,
is a single entity.

Expanded the example to explicitly put the chroot where machinectl can
find it (for subsequent provisioning) and to generate an uploadable
tarball (zstd compression is 10x faster than gzip at a better ratio).