Import or clone an existing image
Plugin is renamed to packer-builder-nspawn to denote that it can now be
used as a generic nspawn container builder with or without debootstrap.

The prepare step will now fail if the named container already exists and
the `-force` option wasn't used.

Example config was rewritten from JSON to HCL2 for easier pipelining.

All container interactions were extracted to Machine class.